requirements.txt
absl-py==2.1.0
altair==5.2.0
appdirs==1.4.4
asttokens==2.4.1
attrs==23.2.0
awscli==1.32.58
blinker==1.7.0
boto3==1.34.58
botocore==1.34.58
cachetools==5.3.3
certifi==2023.7.22
chardet==4.0.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.4
comm==0.2.1
contourpy==1.2.0
cycler==0.10.0
Cython==3.0.9
debugpy==1.8.1
decorator==5.1.1
docker==7.0.0
docker-pycreds==0.4.0
docutils==0.16
executing==2.0.1
filelock==3.13.1
fire==0.5.0
Flask==3.0.2
fonttools==4.49.0
fsspec==2024.2.0
gitdb==4.0.11
GitPython==3.1.42
grpcio==1.62.1
huggingface-hub==0.21.4
idna==2.10
imageio==2.31.1
importlib-metadata==6.11.0
ipykernel==6.29.3
ipython==8.22.2
itsdangerous==2.1.2
jaraco.classes==3.3.1
jedi==0.19.1
Jinja2==3.1.3
jmespath==1.0.1
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter_client==8.6.0
jupyter_core==5.7.1
keyring==8.7
keyrings.alt==5.0.0
kiwisolver==1.4.5
lapx==0.5.7
Markdown==3.5.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.3
matplotlib-inline==0.1.6
mdurl==0.1.2
more-itertools==10.2.0
mpmath==1.3.0
nest-asyncio==1.6.0
networkx==3.2.1
numpy==1.26.4
opencv-python==4.7.0.72
opencv-python-headless==4.8.0.74
packaging==23.2
pandas==2.2.1
parso==0.8.3
Pillow==9.5.0
platformdirs==4.2.0
prompt-toolkit==3.0.43
protobuf==4.21.2
psutil==5.9.8
pure-eval==0.2.2
py-cpuinfo==9.0.0
pyarrow==15.0.1
pyasn1==0.5.1
pybboxes==0.1.6
pydeck==0.8.0
Pygments==2.17.2
Pympler==1.0.1
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-magic==0.4.27
pytube==15.0.0
pytz==2024.1
pytz-deprecation-shim==0.1.0.post0
pywin32==306
pywin32-ctypes==0.2.2
PyYAML==6.0.1
pyzmq==25.1.2
referencing==0.33.0
requests==2.31.0
requests-toolbelt==1.0.0
rich==13.7.1
roboflow==1.1.22
rpds-py==0.18.0
rsa==4.7.2
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
s3transfer==0.10.0
sahi==0.11.15
scipy==1.12.0
seaborn==0.13.2
sentry-sdk==1.41.0
setproctitle==1.3.3
setuptools==69.1.1
shapely==2.0.3
six==1.16.0
smmap==5.0.1
stack-data==0.6.3
streamlit==1.32.2
supervision==0.2.1
sympy==1.12
tabulate==0.9.0
tenacity==8.2.3
tensorboard==2.16.2
tensorboard-data-server==0.7.2
termcolor==2.4.0
terminaltables==3.1.10
thop==0.1.1.post2209072238
toml==0.10.2
toolz==0.12.1
torch==2.2.1
torchvision==0.17.1
tornado==6.4
tqdm==4.66.2
traitlets==5.14.1
typing_extensions==4.10.0
tzdata==2024.1
tzlocal==4.3.1
ultralytics==8.0.196
urllib3==2.0.7
validators==0.22.0
voluptuous==0.14.2
watchdog==4.0.0
wcwidth==0.2.13
Werkzeug==3.0.1
yolo==0.3.1
yolov8==0.0.2
zipp==3.17.0